requirements.txt
apache-airflow==2.10.0
dbt-core==1.8.5
dbt-timescaledb==1.8.0b2
dlt==0.5.3
dlt[postgres]>=0.5.1
mongoengine==0.28.2
pymongo>=4.3.3
pymongoarrow>=1.3.0

